debug: use sync functions

+++ index.js
@@ -213,12 +213,16 @@ async function installPlugins(helm) {
     // so we clean up manually here
     
     let clone_dir = pluginDir + plugin.url.trim().replace(/\/+$/g, '').replace(/[:/]+/g, '-');
-    fs.access(clone_dir, fs.constants.F_OK, (err) => {
-      console.log(`${clone_dir} ${err ? 'does not exist' : 'exists'}`);
-      if(!err) {
-        fs.rmSync(clone_dir, { recursive: true})
-      }
-    });
+    let exists = false;
+    try {
+      fs.accessSync(clone_dir, fs.constants.F_OK);
+      core.debug(`${clone_dir} 'exists'}`);
+      exists = true;
+    }
+    catch(e) {
+      core.debug(`${clone_dir} 'does not exist'}`);
+    }
+    if(exists) fs.rmSync(clone_dir, { recursive: true});
   }
 }
